+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `wait` and `iter_wait` support `log_interval` and `log_level` for printing
   the thrown `PredicateNotSatisfied` to the log.
 - `takesome`: a new generator that partially yields a sequence
+- `repr` and `hash` to typed struct fields.
 
 ### Fixed
 - `ExponentialBackoff`: return the value **before** the incrementation.